requirements.txt
alpaca-py==0.28.0
pandas==2.2.2
python-dotenv==1.0.1